Separated script control functions.
