+++ lib/script_controller.hpp
@@ -1,0 +1,19 @@
+// Copyright (c) 2010-2025, Lawrence Livermore National Security, LLC. Produced
+// at the Lawrence Livermore National Laboratory. All Rights reserved. See files
+// LICENSE and NOTICE for details. LLNL-CODE-443271.
+//
+// This file is part of the GLVis visualization tool and library. For more
+// information and source code availability see https://glvis.org.
+//
+// GLVis is free software; you can redistribute it and/or modify it under the
+// terms of the BSD-3 license. We welcome feedback and contributions, see file
+// CONTRIBUTING.md for details.
+
+#ifndef GLVIS_SCRIPT_CONTROLLER_HPP
+#define GLVIS_SCRIPT_CONTROLLER_HPP
+
+#include <iostream>
+
+void PlayScript(std::istream &scr);
+
+#endif // GLVIS_SCRIPT_CONTROLLER_HPP
